preparation part 4 bugfix

+++ src/main/java/org/example/notificationsbot/model/Notification.java
@@ -22,7 +22,7 @@ public class Notification extends AbstractEntity {
     Status status;
 
     @Column(name = "seconds")
-    Long seconds;
+    Integer seconds;
 
     @ManyToOne
     @JoinColumn(name = "user_id")
